Use x86_64-unknown-uefi target
Rust now supports uefi as a build target: https://github.com/rust-lang/rust/pull/56769

This change mainly eliminates references to the custom JSON target
file. It also requires that the entry point's name be changed to
`efi_main`.

Note that the "C" abi is now correct for EFI applications, but will
still be incorrect when uefi-rs is brought in as an external dependancy
for a different target. This means the entry point should always be
`extern "C"` while the table of function pointers should be `extern
"win64"`.