Add shim lock protocol
This protocol is not part of the UEFI specification, but is installed
by the [Shim bootloader](https://github.com/rhboot/shim) which is
commonly used by Linux distributions to support UEFI Secure Boot. Shim
is built with an embedded certificate that is used to validate another
EFI application before running it. That application may itself be a
bootloader that needs to validate another EFI application before
running it, and the shim lock protocol exists to support that.