Rename boot services' memset to set_mem
In boot.rs we see there is an unsafe function named memset. This function has a signature different than memset(3), but has a name different than that in the edk2 sources. This is confusing and will lead to bugs in a very unsafe function when people mix up the order of the value and the size.

We should either name the function memset and give it the libc-style signature memset(&self, buffer: *mut u8, value: u8, size: usize), or name the function set_mem and give it the signature from EFI_SET_MEM.

uefi-rs/src/table/boot.rs

Line 529 in 948464c

 pub unsafe fn memset(&self, buffer: *mut u8, size: usize, value: u8) { 

See https://github.com/rust-osdev/uefi-rs/issues/232