Implement image loading/starting
This adds code to load, unload, and start images, as well as set the
command-line of the image.

I didn't add anything to uefi-test-runner for this since to really
test it you need a valid EFI binary to load, but I tested it in a
separate project and verified it can load a Linux kernel that has the
EFI stub built in.

A couple things are left unimplemented here:

- LoadImage supports loading from either a DevicePath or from a
  buffer. Only the latter method is implemented here.

- StartImage can provide exit data if the called image exits. I
  haven't implemented passing that back to the caller for now since I
  don't have a use case to test it against.