device path: add header struct
All device paths start with the same header containing type, subtype, and
length. Move these fields into a struct and use the struct in the two existing
device path types, `DevicePath` and `AcpiDevicePath`.

This also fixes the missing `length` field in `AcpiDevicePath` since it uses
the header type now.

The `header` fields in `DevicePath` and `AcpiDevicePath` are not public
because it's generally unsafe to modify the values since they determine how
the rest of the data in the node is interpreted.