device path: add iter method and test
Device paths are variable length structures packed into an array, so usually
you have a pointer to the head of the array and advance through the elements by
adding the length from the header to the pointer. The new `DevicePath::iter`
method provides a more convenient way of doing this.