Add CString16
For now the API is quite minimal; mostly it just makes interop between
regular Rust strings and EFI strings easier. You can convert from a
`&str` to a `CString16`, and since it derefs to `&CStr16` you can call
the methods exposed by that type as well.

This type is gated by the `exts` feature since it requires allocation.