Add CStr16::from_str_with_buf
This provides an easy way to convert from a UTF-8 str to a CStr16
without requiring allocation. You just have to provide a backing buffer
large enough to store the converted characters.

Also converted the runtime vars test to use this method instead of
`CString16`, since if the runtime tests are ever moved to run after
exiting boot services, they won't be able to use allocation.