Fix flaky screenshot test
When running the tests under x86_64 OVMF, the screenshot test sometimes
fails with a timeout error when attempting to read from the serial
port. This timeout occurs very quickly, even though the serial timeout
has been set to 10 seconds.

This behavior is caused by OVMF's implementation of the Simple Text
Input protocol: it periodically polls for input to the serial device,
and changes the serial device's timeout to a very short value prior to
doing so. (See
edk2/MdeModulePkg/Universal/Console/TerminalDxe/TerminalConIn.c.) If
that change to the timeout happens to occur between where
uefi-test-runner sets the timeout and does the serial read, the read may
only wait for a couple hundred microseconds before timing out.

Since this is a timing-related bug it can be hard to reproduce. It seems
to happen much less frequently in headless mode, so it only occasionally
flakes in the Github CI, but happens much more often when running
without `--headless` in a dev environment.

Fix by retrying the serial read a few times if it fails with a timeout
error.

Tested by running the proto tests in a loop:

    for _ in 0..200 { proto::test(image, &mut st); }