Fix an uninitialized-memory error
Clippy is failing with this error in `BootServices::find_handles`:

    error: calling `set_len()` immediately after reserving a buffer creates uninitialized values

Fix by using `Handle::uninitialized`, which uses `MaybeUninit`
internally. This may not be quite as efficient since the vec data is
getting initialized now, but it seems unlikely to matter in practice.

Also replaced a second unsafe call to `set_len` with `truncate`.