Implement OpenProtocol and CloseProtocol in BootServices
OpenProtocol is an extended version of HandleProtocol. According to the
UEFI spec, new code should use OpenProtocol instead of HandleProtocol,
so updated the docstrings to indicate that.

Implemented OpenProtocol as two methods: `open_protocol` and
test_protocol. The latter method corresponds to the
EFI_OPEN_PROTOCOL_TEST_PROTOCOL attribute; it just tests if a protocol
exists on a handle but doesn't open it. The `open_protocol` method
returns a new `ScopedProtocol` that automatically closes the protocol on
drop. Since `open_protocol` takes three different `Handle`s as input,
added an `OpenProtocolParams` struct to clearly name the input `Handle`s
at call sites.