Fix flaky screenshot test
When running the tests under x86_64 OVMF, the screenshot test sometimes
fails with a timeout error when attempting to read from the serial
port. This timeout occurs very quickly, even though the serial timeout
has been set to 10 seconds.

This behavior is caused by OVMF's implementation of the Simple Text
Input protocol: it periodically polls for input to the serial device,
and changes the serial device's timeout to a very short value prior to
doing so. If that change to the timeout happens to occur between where
uefi-test-runner sets the timeout and does the serial read, the read may
only wait for a couple hundred microseconds before timing out.

As explained on the edk2 mailing list, the fix is to open the serial
protocol on the serial handle in `Exclusive` mode so that the simple
text input protocol is disconnected from the handle:
https://edk2.groups.io/g/devel/message/83921