build.py: enable uefi package features
Features are checked against the current package, which by default is
detected by the CWD. This means that if you are in the test runner,
enabling the "exts" feature doesn't work because that package doesn't
have such a feature, only the uefi packages does. So, as with the
`build.py test` command, set the manifest path to the uefi crate so that
this command does the right thing regardless of the CWD.

Also changed "--all" to "--workspace", since the former is deprecated.