Fix logs cut off after screenshot test
OVMF sets up the stdout handle such that writes go to all serial devices
(as well as the visual console when not running headless). This is
convenient since the logger writes to stdout, so connecting a serial
device to QEMU's stdout writes logs from the test runner to the host.

However, once a serial device is opened in exclusive mode it is
disconnected from stdout. Since the test runner needs to open a serial
device in exclusive mode, a single serial device isn't sufficient --
once the device is opened in exclusive mode it will stop receiving logs.

So, open two serial devices for the QEMU test. The first one is
connected to the host's stdout, and serves just to transport logs. The
second one is connected to a pipe, and used to receive the SCREENSHOT
command and send the response. That second will also receive logs up
until the test runner opens the handle in exclusive mode, but we can
just read and ignore those lines.

Fixes https://github.com/rust-osdev/uefi-rs/issues/327