publishing: drop recommendation to manually update deps
Manually updating deps shouldn't be needed since this repo has
dependabot enabled.

Note that the currently-configured behavior of dependabot is slightly
different from the default behavior of running `cargo upgrade`, but I
think in a good way. Specifically, `cargo upgrade` will change deps to
the latest release, whereas dependabot will only change the dep if the
current version specification isn't compatible with latest version of
the dep.

To give a concrete example, uefi-macros currently has this dependency:
`trybuild = "1.0.53"`. That is equivalent to `trybuild = "^1.0.53"`,
meaning that any version satisfying `>=1.0.53 && <2.0.0` is allowed by
the dependency resolver. The latest version of `trybuild` is currently
`1.0.54`, which is compatible with the above version range, so
dependabot won't update it. On the other hand, running `cargo upgrade`
would change the version. In general this is fine, except that if an
end-user's application happens to pull in another dependency that for
whatever reason is compatible with trybuild-1.0.53, but not
trybuild-1.0.54, the resolver would now have to pull in both versions to
satisfy the constraints.