Don't run doctests with invalid pointers
The `Handle::from_ptr` and `SystemTable<View>::from_ptr` doctests use a
`0xdeadbeef` pointer, which of course is not valid. Change the tests to
compile but not actually run so that they don't trigger UB.

Noticed this when running the unit tests with Miri, which caught the
undefined behavior the in the SystemTable case.