Add BootServices::load_image
This replaces `BootServices::load_image_from_buffer` with a single
method that can load from either a `[u8]` buffer or a file path via
`SimpleFileSystem`.

The source is modeled with the `LoadImageSource` enum. That helps make
it clear that the buffer input isn't needed when specifying a file path,
but the reverse isn't necessarily true since the file path can be added
as metadata when loading from a buffer. Additionally, the `boot_policy`
flag (called `from_boot_manager` in `LoadImageSource`) is only used when
loading from a file path.