Expand file system tests
Add code in xtask to build a disk image with a FAT partition, then run
file system tests against that disk (when the test runner is built with
the qemu feature). A disk image is used instead of QEMU's VVFAT option
so that everything is fully controlled (can test exact FS info values)
and isolated from the host.

This partially addresses
https://github.com/rust-osdev/uefi-rs/issues/57. There are some more
partition and file system tests that can be added here.