Add UnalignedCStr16
This wrapper type can be used with UEFI strings that are inside a
`repr(packed)` struct. Creating a reference to a packed field is not
allowed because it might not be properly aligned, so a `CStr16` can't be
directly constructed. `UnalignedCStr16` instead takes a pointer, which
is allowed. The resulting unaligned string cannot be used directly, but
must be converted to an aligned form first with `to_cstr16` (requires a
backing buffer) or `to_cstring16` (requires the `exts` feature).