Add FilePathMediaDevicePath and DevicePath conversion methods
`FilePathMediaDevicePath` is similar to `AcpiDevicePath`, but for
type=MEDIA/subtype=MEDIA_FILE_PATH. It is a DST because it contains a
variable-length string at the end. Since it is also a packed struct, the
path must be accessed via `UnalignedCStr16`.

Also added `DevicePath::as_acpi_device_path` and
`DevicePath::as_file_path_media_device_path` to make it easy to safely
convert a `DevicePath` node to a more specific type. (There are many
more `DevicePath` node types yet to be added though.)

https://github.com/rust-osdev/uefi-rs/issues/401