Improve device path API
This is an update to the device path API with two goals:
* Clearly define the separation between device paths, device path
  instances, and device path nodes by using different types.
* Make tests pass under Miri with `-Zmiri-tag-raw-pointers` enabled

In the UEFI spec, the `EFI_DEVICE_PATH_PROTOCOL` struct is used to
represent both full paths and individual nodes. Prior to this commit,
`DevicePath` was used the same way. This didn't make for the clearest
API though, since the distinction between paths and nodes mainly existed
in the documentation rather than the type system. To solve this, there
are now three main types: `DevicePath`, `DevicePathInstance`, and
`DevicePathNode`.

In addition, since `DevicePath` was only four bytes in size (i.e. the
size of the node header), Miri was unhappy when bytes were read past
those four bytes. The solution is to use DSTs that cover the appropriate
span of bytes. This required the addition of a new `ProtocolPointer`
trait so that when a DST protocol is opened, a fat pointer can be
constructed from the thin `c_void` pointer.

The `device_path` module's docstring has been expanded quite a bit to
describe how all the types fit together.

This commit enables `-Zmiri-tag-raw-pointers`, so:
Fixes https://github.com/rust-osdev/uefi-rs/issues/414