Switch back to automatic Debug derive for Header struct
The `Header` `Debug` implementation was accidentally printing the `size`
field for the `signature`. Rather than just fix that field, switch the
whole thing back to using `derive(Debug)`. This adds in the `_reserved`
field, which probably doesn't matter much either way but might as well
show the whole struct when debug printing.