Fix compilation of the xtask package under Windows
The qemu code makes use of some APIs from the `nix` crate, which causes
compilation to fail on Windows (native, not WSL). The qemu code can't be
easily made to work on Windows, but the rest of xtask should still work
fine building, running clippy, etc. Fix by adding `#![cfg(unix)]` to the
`qemu` module, as well as the `disk` and `net` modules that are only
used by the qemu code. Also changed the `run_vm_tests` action to panic
on non-Unix targets.