Fix query_variable_info test
The behavior of OVMF changed recently so that calling
query_variable_info with just the `NON_VOLATILE` attribute is treated as
an error:
https://github.com/tianocore/edk2/commit/21320ef66989f8af5a9e9b57df73d20a70bea85f

Fix by OR'ing it with `BOOTSERVICE_ACCESS`.

Also fix a typo in the output, and `unwrap` the result of calling
`query_variable_info` since these calls should always succeed.