Simplify uefi-services panic handler
The `PanicInfo` type implements `Display`, so it can be printed out in a
standard format without manually accessing the location and message data.

This changes the format of the panic output a bit, and now matches the
"normal" panic output of an std Rust target.

Old output:

```
[ERROR]: uefi-services/src/lib.rs@144: Panic in uefi-test-runner/src/main.rs at (44, 5):
[ERROR]: uefi-services/src/lib.rs@151: test panic
```

New output:

```
[ERROR]: uefi-services/src/lib.rs@141: panicked at 'test panic', uefi-test-runner/src/main.rs:44:5
```

This also allows the `panic_info_message` unstable feature to be removed.