Add MemoryProtection protocol
This corresponds to the `EFI_MEMORY_ATTRIBUTE_PROTOCOL` proposal:
https://bugzilla.tianocore.org/show_bug.cgi?id=3519

The protocol is part of a push to add better memory protection to UEFI:
https://github.com/MicrosoftDocs/windows-driver-docs/blob/abb7dfa01467a300eefe589a4a8f9353199c0ade/windows-driver-docs-pr/bringup/uefi-ca-memory-mitigation-requirements.md

This is not yet part of a released UEFI Specification, but support for
using the protocol has already been added to shim and Fedora's fork of grub
* https://github.com/rhboot/shim/pull/459/commits/825d99361b4aaa16144392dc6cea43e24c8472ae
* https://github.com/rhboot/grub2/commit/aa3c4f62715c81de280de8403d0de1ff52a7440c

By our usual naming conventions, `EFI_MEMORY_ATTRIBUTE_PROTOCOL` would
translate to a struct named `MemoryAttribute`. I went with
`MemoryProtection` instead because:

1. We don't typically include `Protocol` in the name of our `Protocol`
   structs, but there's already a `MemoryAttribute` type in the crate,
   which is the bitflag enum that describes the actual attributes. Since
   the two types are used together, the naming conflict would be
   confusing.
2. This protocol only gets and sets the RO/RP/XP memory attribute flags,
   so the generic `EFI_MEMORY_ATTRIBUTE_PROTOCOL` name is a bit
   misleading; there are many other memory attributes that it does not
   cover.

I put the new protocol under `proto::security` since the proposal would
place the new protocol in the UEFI Specification under section 37
"Secure Technologies"; `proto::secure_technologies` is a bit verbose so
went with `proto::security` instead.