ci: Deploy book to gh-pages
This workflow is triggered by pushes to `main`. It will update the
`gh-pages` branch, throwing out any prior history to avoid increasing
the overall size of repo unnecessarily.

For now we just build `main`, but in the future it might be good to
build a `stable` version as well. To avoid breaking existing URLs, the
book is placed under the `HEAD` subdirectory, with a redirect page in
the root.