xtask: Add ChildWrapper
This wraps the QEMU child process so that the process is killed on
drop. This avoids leaving a dangling qemu behind if an unexpected error
happens during the test.

Also add a context message to the spawn error, makes it easier to see
what's wrong if for example qemu isn't in $PATH.