xtask: Enable QEMU under Windows without WSL
This updates the `Pipe` struct to work with the Windows version of named
pipes as well as Unix pipes. The `Io` struct has also been modified a
little bit so that JSON can be sent without a trailing newline, as that
causes QEMU to hang on Windows.

Fixes https://github.com/rust-osdev/uefi-rs/issues/461