xtask: Improve echo service lifecycle
If the program ends while the echo service is still blocked in
`recv_from`, it panics on Windows:

```
thread '<unnamed>' panicked at 'failed to receive packet: Os { code: 10004, kind: Uncategorized, message: "A blocking operation was interrupted by a call to WSACancelBlockingCall." }', xtask\src\net.rs:17:14
```

Fix by using adding a timeout to the reads so that a shutdown flag can
be checked. Also wrap the service in a struct to ensure join on drop.