Allow qemu test to work under Windows without WSL (#463)
* xtask: Add ChildWrapper

This wraps the QEMU child process so that the process is killed on
drop. This avoids leaving a dangling qemu behind if an unexpected error
happens during the test.

Also add a context message to the spawn error, makes it easier to see
what's wrong if for example qemu isn't in $PATH.

* xtask: Improve echo service lifecycle

If the program ends while the echo service is still blocked in
`recv_from`, it panics on Windows:

```
thread '<unnamed>' panicked at 'failed to receive packet: Os { code: 10004, kind: Uncategorized, message: "A blocking operation was interrupted by a call to WSACancelBlockingCall." }', xtask\src\net.rs:17:14
```

Fix by using adding a timeout to the reads so that a shutdown flag can
be checked. Also wrap the service in a struct to ensure join on drop.

* xtask: Add platform module

This module contains functions for determining the host platform. These
are intended to be used instead of `#[cfg(...)]` blocks. Such blocks can
hide a compilation problem on a different platform, whereas using these
functions will allow stuff to be compiled for all platforms.

* xtask: Restrict KVM to Linux

KVM is a Linux feature, as far as I know not available on any other OSes.

* xtask: Enable QEMU under Windows without WSL

This updates the `Pipe` struct to work with the Windows version of named
pipes as well as Unix pipes. The `Io` struct has also been modified a
little bit so that JSON can be sent without a trailing newline, as that
causes QEMU to hang on Windows.

Fixes https://github.com/rust-osdev/uefi-rs/issues/461