xtask: Always allow writing to OVMF vars
In the AArch64 version of OVMF currently used in CI, it crashes if the
vars aren't writable. (A newer version that I tested locally doesn't
have this problem.) Having vars be writable is a more normal UEFI
environment anyway though, so remove the special case for AArch64 and
make the vars always writable.

Since the vars file might be in some read-only location, or owned by
root (e.g. a system package installed to /usr), a temporary copy of the
vars file is made before using it.

Also add a `PflashMode` enum to make the call sites of `add_pflash_args`
a little clearer.