xtask: Add default QEMU executable directory on Windows
The QEMU installer for Windows does not automatically add the directory
containing the QEMU executables to the PATH. Add the default directory
to the PATH to make it more likely that QEMU will be found on
Windows. The directory is appended, so if a different directory on the
PATH already has the QEMU binary this change won't affect anything.