Pin to a working nightly
Temporarily modify rust-toolchain.toml to pin to `nightly-2022-08-26`,
as compilation is broken on current nightly due to
https://github.com/rust-osdev/uefi-rs/issues/500. `nightly-2022-09-01`
is the most recent nightly that avoids the bug, but it doesn't have miri
(which is run by CI), so pin to a slightly older version.

Also pinned the toolchain in the template app.

This change should be reverted once #500 is fixed.