Temporarily revert use of core::ffi::CStr
The move of `CStr` from `std` to `core` is not due to become stable
until 1.64. The first nightly where it could be used without the
unstable `core_c_str` feature was 2022-07-16, which is still pretty
recent. Let's back this change out for now to give users pinned to older
versions more time.