Add initial support for the TCG protocols
The TCG protocols are for interacting with TPMs, or Trusted Platform
Modules. There are two versions in common use: TPM 1.2 and TPM 2.0. The
TCG protocol is for TPM 1.x, and the TCG2 protocol can theoretically be
used for any TPM but in practice is usually implemented just for TPM
2.0.

To avoid dumping too much into one PR, this commit is an MVP that only
includes getting information from the TPM (`status_check` for v1,
`get_capability` for v2).

I've exposed the new protocols with the same name but in different
module paths:
```
uefi::proto::tcg::v1::Tcg
uefi::proto::tcg::v2::Tcg
```

There is a lot of overlap between the two protocols, and the TPM v2
protocol relies in some places on v1 (although those parts are not
included in this MVP). Also, it is common to need to support both
versions at the same time, so I expect a common usage pattern will be to
import `uefi::proto::tcg::{v1, v2}` and refer to TCG types with
`v1::...` and `v2::...`.