ci: Test the build on an old nightly
Add a CI job that runs `cargo xtask build` with an old
nightly. Currently the oldest nightly that works is 2022-04-18. There's
nothing particularly magic about that version; we can require a newer
version when we deem it appropriate. Having this CI job just makes it
easy for us to identify in advance when a PR is going to require a newer
nightly, while still allowing other CI jobs to use latest nightly.

This change is related to (but not a fix for):
https://github.com/rust-osdev/uefi-rs/issues/510