Add PhysicalAddress and VirtualAddress type aliases
These correspond to `EFI_PHYSICAL_ADDRESS` and
`EFI_VIRTUAL_ADDRESS`. These usually end up being converted to a pointer
type, but notably they are always 64-bit regardless of the target
platform. This change revealed a bug in the `AllocateType` enum; it was
using `usize` for a couple members, but that is not correct for all
targets, e.g. `i686-unknown-uefi`.