Add option to disable network tests
`cargo xtask run` now takes a `--disable-network` arg that turns off the
network protocol tests by not creating a network device.

The purpose of this flag is to make the QEMU tests faster when iterating
on something locally. Disabling the network stuff drops the time down
from ~12.5s to ~4.5s. The extra time doesn't matter when running in CI,
but for quick local testing it can add up.