ci: Simplify some jobs with ubuntu-22.04 runner
The aarch64 and ia32 jobs had some extra complexity to get a working
OVMF on ubuntu-20.04. Now that there are ubuntu-22.04 runners available,
switch the jobs to that version and use the provided OVMF packages.

Eventually "ubuntu-latest" will point to ubuntu-22.04 instead of
ubuntu-20.04, but for now it must be specified explicitly.