Run the ia32 target using the same settings as x86_64
This fixes `cargo xtask run --target ia32` when using the `ovmf-ia32`
Ubuntu package. The important difference is really `-machine q35`, but
we might as well share all the settings.

Running the 32-bit firmware on a 64-bit machine is actually a slightly
more realistic test than running it on a 32-bit machine, since (at least
to my knowledge) ia32 UEFI on a 64-bit CPU is more commonly found than
on a 32-bit CPU.