Change DeviceSubType literals to the same numeric base as the spec
The UEFI Specification uses base 10 for almost all the sub types (the
exception is base 16 for the two `END` node types). Change our enum to
match, as that makes it much easier to see at a glance that the value is
correct.

There should be no change in the actual values from this commit, just
makes the code easier to verify against the spec.