Fix clippy warnings
Clippy now warns when arguments in a format string could be specified
inline, e.g. `format!({var})` instead of `format!({}, var)`. Note that
only standalone variables can be used in this way, e.g. `{var}` works,
but not `{self.var}`.

I think it helps readability to be consistent within a format string and
only use one style, so in cases where any of the arguments could be done
in the new style I introduced new local variables as needed.