Add UnalignedSlice
This is a generalized version of `UnalignedCStr16`. It's similar to
`&[T]`, except it can point to an unaligned slice. This is useful with
device path nodes which sometimes end with a DST slice. Since device
path nodes are all packed structures, the DST slice may be unaligned.

`UnalignedSlice` tries to generally follow the API of a regular slice,
with some deviations where necessary (for example it returns copies
rather than references, and provides a `copy_to` method instead of
`copy_from_slice`).