test-runner: Drop the second serial device
The screenshot test sends a command to the host via the serial device,
then waits for a reply before continuing. We used a second serial device
just for this request and reply so that we could open the protocol in
exclusive mode without breaking logging for tests after that.

It turns out there's a simpler way to do this: we can use the same
serial device as the one used for logs, and after we're finished use
`connect_controller` to restore the connection between the serial device
and the console output.