uefi: Add "unstable" feature
The unstable feature will be used to gate the use of new unstable
features. We hope to make the uefi crate work on the stable channel once
the `abi_efiapi` and `default_alloc_error_handler` features are
stabilized, so we don't want to start relying unconditionally on new
unstable features.

For now the new feature just gates `error_in_core`, but some additional
features will be gated here soon.