test-runner: Open serial device in exclusive mode
The serial test is occasionally failing in CI. I'm hoping that changing
the test to open in exclusive mode will fix it, using the same trick of
reconnecting all controllers that we do in `send_request_to_host`. It's
hard to be sure though since the test failure doesn't occur all that often.

Also updated the comment for skipping the test on aarch64: QEMU+EDK2
doesn't support the software loopback used by the test.