uefi-macros: Add `unsafe_protocol` macro
This attribute macro combines the functionality of the `unsafe_guid`
attribute macro and the `Protocol` derive macro. It impls the `Protocol`
and `Identify` traits, and it also marks the type as `!Send` and
`!Sync`.

This provides a slightly nicer API for protocols (`unsafe_protocol` is a
clearer name than `unsafe_guid`, and only one macro is needed instead of
two), but more importantly it does the `!Send` and `!Sync` part via a
hidden `PhantomData` field rather than relying on the unstable
`negative_impls` feature.