uefi-macros: Refactor GUID tests
Switch the tests from using `unsafe_guid!` to `guid!`. Both use the same
internal code, so we're testing more or less the same thing, but `guid!`
is a little more direct.

Also, as we previously did with the entry macro, split the GUID tests
into multiple files to make it easier to visually inspect the error
output.

The "good" test case has been removed since that's already covered in
plenty of places elsewhere.