uefi: pxe: Fix BaseCode::discover optional argument, tweak DiscoverInfo type.
NOTE: This commit changes the API for `BaseCode::discover`, but should continue
to be source-compatible.

The `info` parameter in `BaseCode::discover` is an optional pointer argument
(it may be NULL), but is incorrectly specified in the Rust exern definition
as `Option<*const T>`. This type is (at least) one byte larger than a Pointer
on all platforms, making it incompatible with the Optional Pointers in EFI.

The `DiscoverInfo` struct had odd bounds associated with it, and being generic
over `T: ?Sized` doesn't make sense when the server list has to be a fixed size.
This could be removed or broken out into a separate commit if needed.