Fix warnings from `abi_efiapi` stabilization
The `abi_efiapi` feature has been stabilized, so in the very latest nightlies we
get this warning from enabling it:

    the feature `abi_efiapi` has been stable since 1.68.0-nightly and no longer requires an attribute to enable

We don't want to stop using `#![feature(abi_efiapi)]` just yet though, because
that would require raising our nightly MSRV to 2023-01-13. So for now, to get
the CI passing again, turn off the `stable_features` warning.

In the `uefi-macros` tests, just remove `#![feature(abi_efiapi)]` since that
code is only run when developing uefi-rs, not when using it.